refactor(yjs): split up into multiple files
